requirements.txt
mcp>=1.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
uvicorn>=0.23.0
